proper Colour type!
- finally got sick of random uint32s infecting function signatures across the engine
- having a defined Colour type allows us to define better convenience functions, which will be important once we start messing around with colour blending and manipulation in other colour spaces
- this also paves the way for maybe defining other types for colours in different formats. i doubt this will ever happen but who knows.

+++ gfx/export.go
@@ -12,7 +12,7 @@ func (c Canvas) ExportToXP(path string) {
 
 	for i, cell := range c.cells {
 		cellData := reximage.CellData{}
-		cellData.SetColoursARGB(cell.Colours.Fore, cell.Colours.Back)
+		cellData.SetColoursARGB(uint32(cell.Colours.Fore), uint32(cell.Colours.Back))
 		switch cell.Mode {
 		case DRAW_GLYPH:
 			cellData.Glyph = uint32(cell.Glyph)
